Added shooter subsystem and formatted code

+++ src/Subsystems/BallIntake.cpp
@@ -1,14 +1,14 @@
 #include "BallIntake.h"
 #include "../RobotMap.h"
 
-BallIntake::BallIntake() : Subsystem("BallIntake") {
+BallIntake::BallIntake():Subsystem("BallIntake"){
 	_IntakeMotor.reset(new frc::Talon(0));
 }
 
-void BallIntake::InitDefaultCommand() {
+void BallIntake::InitDefaultCommand(){
 
 }
 
-void BallIntake::Set(float Power){
+void BallIntake::Set(double Power){
 	_IntakeMotor->Set(Power);
 }
