+++ src/Subsystems/Climber.cpp
@@ -1,14 +1,14 @@
 #include "Climber.h"
 #include "../RobotMap.h"
 
-Climber::Climber() : Subsystem("Climber") {
+Climber::Climber():Subsystem("Climber"){
 	_ClimberMotor.reset(new frc::Talon(1));
 }
 
-void Climber::InitDefaultCommand() {
+void Climber::InitDefaultCommand(){
 
 }
 
-void Climber::Set(float Power){
+void Climber::Set(double Power){
 	_ClimberMotor->Set(Power);
 }
